Перенос api на TypeScript
