+++ CHANGELOG.md
@@ -358,3 +358,4 @@
 - Dockerfile переносит каталог `dist` в образ для корректной работы pm2
 - Переведены на TypeScript сервис, контроллер и роут карт
 - Расширена проверка коротких ссылок Google Maps: DNS-разрешение и фильтрация приватных IP защищают от SSRF
+- API, middleware и swagger переписаны на TypeScript, исходные `.js` удалены
