+++ ROADMAP.md
@@ -284,3 +284,4 @@
 251. Проверяется домен коротких ссылок Google Maps для защиты от SSRF
 252. Проверяется протокол https, userinfo и порт коротких ссылок Google Maps
 253. Уточнена типизация `username` в модуле авторизации для сборки Docker
+254. API и middleware переписаны на TypeScript, исходные `.js` файлы удалены
