+++ bot/src/server.ts
@@ -1,4 +1,4 @@
 // Назначение файла: стартовый скрипт API.
 // Основные модули: container, api
 import './container';
-import './api/api.js';
+import './api/api';
